+++ src/main/kotlin/lotto/StringHandler.kt
@@ -1,0 +1,8 @@
+package lotto
+
+fun tokenizeWinningNumbers(input: String): List<Int> {
+    return input
+        .split(",")
+        .filterNot { it.isBlank() }
+        .map { it.trim().toInt() }
+}
