--- src/main/kotlin/lotto/util/StringHandler.kt
@@ -1,10 +1,0 @@
-package lotto.util
-
-class StringHandler {
-    fun tokenizeWinningNumbers(input: String): List<Int> {
-        return input
-            .split(",")
-            .filterNot { it.isBlank() }
-            .map { it.trim().toInt() }
-    }
-}
