+++ src/main/kotlin/lotto/view/Input.kt
@@ -1,6 +1,6 @@
-package lotto.ui
+package lotto.view
 
-class InputView {
+class Input {
     fun getPurchaseAmount(): String {
         println("구입금액을 입력해 주세요.")
         return readln()
