+++ src/test/kotlin/lotto/domain/PrizeTest.kt
@@ -1,7 +1,5 @@
-package lotto.util
+package lotto.domain
 
-import lotto.Lotto
-import lotto.Lottos
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.params.ParameterizedTest
